Small corrections to get a first version of the configuration working
svn path=/trunk/; revision=6456

+++ sys/win32/msdev/src/plstub.cpp
@@ -631,7 +631,7 @@ plsfnam(const char *fnam) {
 }
 
 /* Shade region. */
-typedef int  DefineFuncPtr( PLFLT, PLFLT ) ;
+typedef long  DefineFuncPtr( PLFLT, PLFLT ) ;
 typedef void *FillFuncPtr(PLINT, PLFLT *, PLFLT *) ;
 typedef void *PlotFuncPtr(PLFLT, PLFLT, PLFLT *, PLFLT *, PLPointer) ;
